+++ src/library/openglwrappers.h
@@ -52,6 +52,7 @@ OVERRIDE int glXGetSwapIntervalMESA(void);
 OVERRIDE void glXQueryDrawable(Display * dpy,  GLXDrawable draw,  int attribute,  unsigned int * value);
 
 OVERRIDE GLXContext glXCreateContextAttribsARB (Display *dpy, GLXFBConfig config, GLXContext share_context, Bool direct, const int *attrib_list);
+OVERRIDE void glXDestroyContext(Display * dpy,  GLXContext ctx);
 
 OVERRIDE const GLubyte* glGetString( GLenum name);